一系列小的修正。 修正媒体库不能正确响应文件的添加和删除问题。  在文件目录上投送监视，响应之后须等一下，不然文件不能正常打开。 播放项目的删除，先删除视图里面的，再删除数据。
git-svn-id: svn://172.16.0.3@201 c573b714-58c8-aa40-881b-c130d9d1abad

+++ playerT/mytree.h
@@ -581,7 +581,7 @@ public:
 
 			oldRC=cur->getRectBeforeSize();
 
-			int curW=(bLR?WIDTH(oldRC):HEIGHT(oldRC)) * times;
+			float curW=(bLR?WIDTH(oldRC):HEIGHT(oldRC)) * times;
 
 			if(count!=childs-1)
 			{
